# Reflection Constructor Binding Refactor