--- a/src/Autofac/Core/Activators/Reflection/BoundConstructor.cs
+++ b/src/Autofac/Core/Activators/Reflection/BoundConstructor.cs
@@ -46,33 +46,31 @@ namespace Autofac.Core.Activators.Reflection
         /// <summary>
         /// Initializes a new instance of the <see cref="BoundConstructor"/> class for a successful bind.
         /// </summary>
-        /// <param name="constructor">The constructor.</param>
+        /// <param name="binder">The binder that generated this binding.</param>
         /// <param name="factory">The instance factory.</param>
         /// <param name="valueRetrievers">The set of value-retrieval functions.</param>
         [SuppressMessage("Design", "CA1062:Validate arguments of public methods", Justification = "Validated in constructor.")]
-        public static BoundConstructor ForBindSuccess(ConstructorInfo constructor, Func<object?[], object> factory, Func<object?>[] valueRetrievers)
-            => new BoundConstructor(constructor, factory, valueRetrievers);
+        public static BoundConstructor ForBindSuccess(ConstructorBinder binder, Func<object?[], object> factory, Func<object?>[] valueRetrievers)
+            => new BoundConstructor(binder, factory, valueRetrievers);
 
         /// <summary>
         /// Initializes a new instance of the <see cref="BoundConstructor"/> class, for an unsuccessful bind.
         /// </summary>
-        /// <param name="constructor">The constructor.</param>
-        /// <param name="totalArgumentCount">The total count of arguments to the constructor.</param>
+        /// <param name="binder">The binder that generated this binding.</param>
         /// <param name="firstNonBindableParameter">The first parameter that prevented binding.</param>
-        public static BoundConstructor ForBindFailure(ConstructorInfo constructor, int totalArgumentCount, ParameterInfo firstNonBindableParameter) =>
-            new BoundConstructor(constructor, totalArgumentCount, firstNonBindableParameter);
+        public static BoundConstructor ForBindFailure(ConstructorBinder binder, ParameterInfo firstNonBindableParameter) =>
+            new BoundConstructor(binder, firstNonBindableParameter);
 
         /// <summary>
         /// Initializes a new instance of the <see cref="BoundConstructor"/> class for a successful bind.
         /// </summary>
-        /// <param name="constructor">The constructor.</param>
+        /// <param name="binder">The binder that generated this binding.</param>
         /// <param name="factory">The instance factory.</param>
         /// <param name="valueRetrievers">The set of value-retrieval functions.</param>
-        internal BoundConstructor(ConstructorInfo constructor, Func<object?[], object> factory, Func<object?>[] valueRetrievers)
+        internal BoundConstructor(ConstructorBinder binder, Func<object?[], object> factory, Func<object?>[] valueRetrievers)
         {
             CanInstantiate = true;
-            TargetConstructor = constructor ?? throw new ArgumentNullException(nameof(constructor));
-            ArgumentCount = valueRetrievers.Length;
+            Binder = binder;
             _factory = factory ?? throw new ArgumentNullException(nameof(factory));
             _valueRetrievers = valueRetrievers ?? throw new ArgumentNullException(nameof(valueRetrievers));
             _firstNonBindableParameter = null;
@@ -81,29 +79,32 @@ namespace Autofac.Core.Activators.Reflection
         /// <summary>
         /// Initializes a new instance of the <see cref="BoundConstructor"/> class, for an unsuccessful bind.
         /// </summary>
-        /// <param name="constructor">The constructor.</param>
-        /// <param name="totalArgumentCount">The total count of arguments to the constructor.</param>
+        /// <param name="binder">The binder that generated this binding.</param>
         /// <param name="firstNonBindableParameter">The first parameter that prevented binding.</param>
-        internal BoundConstructor(ConstructorInfo constructor, int totalArgumentCount, ParameterInfo firstNonBindableParameter)
+        internal BoundConstructor(ConstructorBinder binder, ParameterInfo firstNonBindableParameter)
         {
+            Binder = binder;
             CanInstantiate = false;
-            TargetConstructor = constructor ?? throw new ArgumentNullException(nameof(constructor));
-            ArgumentCount = totalArgumentCount;
             _firstNonBindableParameter = firstNonBindableParameter ?? throw new ArgumentNullException(nameof(firstNonBindableParameter));
             _factory = null;
             _valueRetrievers = null;
         }
 
         /// <summary>
+        /// Gets the binder that created this binding.
+        /// </summary>
+        public ConstructorBinder Binder { get; }
+
+        /// <summary>
         /// Gets the constructor on the target type. The actual constructor used
         /// might differ, e.g. if using a dynamic proxy.
         /// </summary>
-        public ConstructorInfo TargetConstructor { get; }
+        public ConstructorInfo TargetConstructor => Binder.Constructor;
 
         /// <summary>
         /// Gets the total number of arguments for the bound constructor.
         /// </summary>
-        public int ArgumentCount { get; }
+        public int ArgumentCount => Binder.ParameterCount;
 
         /// <summary>
         /// Gets a value indicating whether the binding is valid.