… src/Autofac/Core/Activators/Reflection/BoundConstructorResources.resx
… src/Autofac/Core/Activators/Reflection/ConstructorParameterBinder.cs
… src/Autofac/Core/Activators/Reflection/ConstructorParameterBinding.cs
… src/Autofac/Core/Activators/Reflection/DefaultConstructorFinder.cs
… src/Autofac/Core/Activators/Reflection/IConstructorSelector.cs
… src/Autofac/Core/Activators/Reflection/MatchingSignatureConstructorSelector.cs
… src/Autofac/Core/Activators/Reflection/MostParametersConstructorSelector.cs
… src/Autofac/Core/Activators/Reflection/ReflectionActivator.cs
… src/Autofac/Core/Activators/Reflection/ReflectionActivatorResources.Designer.cs
… src/Autofac/Core/Activators/Reflection/ReflectionActivatorResources.resx
… src/Autofac/Core/Registration/DefaultRegisteredServicesTracker.cs
… test/Autofac.Test/Core/Activators/Reflection/ConstructorBinderTests.cs
… test/Autofac.Test/Core/Activators/Reflection/MatchingSignatureConstructorSelectorTests.cs
… test/Autofac.Test/Core/Activators/Reflection/MostParametersConstructorSelectorTests.cs
… test/Autofac.Test/Core/Activators/Reflection/ReflectionActivatorTests.cs
… test/Autofac.Test/Factory.cs
… test/Autofac.Test/Mocks.cs