@alistairjevans 2020-06-27
Oof, yeah, that MostParametersConstructorSelector was overdue for some love. Refactor is in, results based on the benchmark pending for develop:

### develop

| Method |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-----------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
| Single |   994.6 ns |  4.53 ns |  4.01 ns |  1.00 |    0.00 | 0.3071 |     - |     - |   1.26 KB |
|    Two | 2,112.8 ns |  8.60 ns |  7.18 ns |  2.12 |    0.01 | 0.6638 |     - |     - |   2.72 KB |
|  Three | 3,022.0 ns |  9.73 ns |  8.13 ns |  3.04 |    0.01 | 0.9499 |     - |     - |   3.88 KB |
|   Four | 3,893.8 ns | 14.52 ns | 12.12 ns |  3.91 |    0.02 | 1.2970 |     - |     - |   5.32 KB |

### constructor-binder

| Method |       Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
| Single |   893.8 ns |  2.50 ns |  1.95 ns |  1.00 | 0.3576 |     - |     - |   1.46 KB |
|    Two | 1,492.8 ns |  6.43 ns |  5.70 ns |  1.67 | 0.5474 |     - |     - |   2.24 KB |
|  Three | 2,182.4 ns |  9.87 ns |  8.75 ns |  2.44 | 0.7820 |     - |     - |    3.2 KB |
|   Four | 2,999.6 ns | 12.39 ns | 10.35 ns |  3.36 | 1.0567 |     - |     - |   4.32 KB |