@alistairjevans 2020-06-27
I've revisited how we allocate bindings, and re-allocate those invalid bindings, as you mentioned.

Where I've ended up is that the `IConstructorSelector` receives the original array of bindings rather than a filtered set. This means that the selectors are now responsible for ignoring invalid bindings.

This is obviously a breaking behaviour change, but the new requirement is a relatively easy one to integrate for that small set of users who have implemented a custom `IConstructorSelector`.

Memory and perf gains for the 'one of my constructors isn't valid' path:


### develop

|             Method |     Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|----------:|----------:|------:|--------:|-------:|------:|------:|----------:|
|             Single | 1.010 us | 0.0069 us | 0.0057 us |  1.00 |    0.00 | 0.3071 |     - |     - |   1.26 KB |
|                Two | 2.153 us | 0.0161 us | 0.0151 us |  2.13 |    0.02 | 0.6638 |     - |     - |   2.72 KB |
| TwoValidOneInvalid | 2.738 us | 0.0108 us | 0.0101 us |  2.71 |    0.02 | 0.8392 |     - |     - |   3.44 KB |
|              Three | 3.041 us | 0.0178 us | 0.0158 us |  3.01 |    0.03 | 0.9499 |     - |     - |   3.88 KB |
|               Four | 3.917 us | 0.0192 us | 0.0170 us |  3.88 |    0.03 | 1.2970 |     - |     - |   5.32 KB |

## constructor-binder with filtering before-hand

|             Method |       Mean |    Error |   StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |-----------:|---------:|---------:|------:|--------:|-------:|------:|------:|----------:|
|             Single |   893.1 ns |  4.59 ns |  4.07 ns |  1.00 |    0.00 | 0.3576 |     - |     - |   1.46 KB |
|                Two | 1,489.1 ns |  9.34 ns |  8.74 ns |  1.67 |    0.01 | 0.5474 |     - |     - |   2.24 KB |
| TwoValidOneInvalid | 2,188.6 ns | 17.92 ns | 16.77 ns |  2.45 |    0.02 | 0.7324 |     - |     - |   2.99 KB |
|              Three | 2,171.2 ns | 14.34 ns | 13.41 ns |  2.43 |    0.02 | 0.7820 |     - |     - |    3.2 KB |
|               Four | 2,975.3 ns | 18.54 ns | 17.35 ns |  3.33 |    0.02 | 1.0567 |     - |     - |   4.32 KB |


### constructor-binder with selectors getting all bind results

|             Method |       Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |-----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
|             Single |   896.5 ns |  4.61 ns |  3.85 ns |  1.00 | 0.3576 |     - |     - |   1.46 KB |
|                Two | 1,511.2 ns |  7.57 ns |  7.08 ns |  1.69 | 0.5474 |     - |     - |   2.24 KB |
| TwoValidOneInvalid | 1,955.7 ns |  8.92 ns |  8.34 ns |  2.18 | 0.6866 |     - |     - |   2.81 KB |
|              Three | 2,187.0 ns |  9.01 ns |  7.99 ns |  2.44 | 0.7820 |     - |     - |    3.2 KB |
|               Four | 2,964.0 ns | 11.71 ns | 10.96 ns |  3.31 | 1.0567 |     - |     - |   4.32 KB |